+++ src/app/components/layouts/column-container/terra-two-columns-container.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+/**
+ * @author mfrank
+ */
+@Component({
+    selector: 'terra-2-col',
+    templateUrl: require('./terra-two-columns-container.component.html')
+})
+export class TerraTwoColumnsContainerComponent
+{}